Divided the ChatViewModel and Repositroy to anothe 2 ViewModels and 3 Repositories without testing
