+++ chat/feat-chat/src/main/java/com/seravian/feat_chat/data/repository/ChatBotState.kt
@@ -1,0 +1,13 @@
+package com.seravian.feat_chat.data.repository
+
+import com.greenvenom.core_network.data.EmptyResult
+import com.greenvenom.core_network.data.NetworkError
+import com.seravian.core_chat.domain.models.Chat
+import com.seravian.core_chat.domain.models.Message
+
+data class ChatBotState(
+    val currentChat: Chat ?= null,
+    val lastMessage: Message ?= null,
+    val isWaitingForResponse: Boolean = true,
+    val joinChatResult: EmptyResult<NetworkError> ?= null
+)
