+++ chat/feat-chat/src/main/java/com/seravian/feat_chat/data/repository/VoiceModeRepositoryImpl.kt
@@ -1,0 +1,87 @@
+package com.seravian.feat_chat.data.repository
+
+import com.greenvenom.core_network.data.EmptyResult
+import com.greenvenom.core_network.data.ErrorType
+import com.greenvenom.core_network.data.NetworkError
+import com.greenvenom.core_network.data.NetworkResult
+import com.greenvenom.core_network.data.map
+import com.seravian.core_chat.data.dto.request.FetchAIAudioRequest
+import com.seravian.core_chat.data.dto.request.UploadVoiceRequest
+import com.seravian.core_chat.domain.models.Audio
+import com.seravian.feat_chat.domain.ChatBotRemoteDataSource
+import com.seravian.feat_chat.domain.repository.VoiceModeRepository
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.async
+import kotlinx.coroutines.sync.Mutex
+
+class VoiceModeRepositoryImpl(
+    private val seravianChatBotDataSource: ChatBotRemoteDataSource
+): VoiceModeRepository {
+    private val audioDownloadLock: Mutex = Mutex()
+    private var downloadExecution: Deferred<NetworkResult<Audio, NetworkError>>? = null
+
+    private var lastNotifiedAudio: Long = -1
+
+    override suspend fun sendCapturedVoice(
+        capturedVoice: ByteArray,
+        chatId: String
+    ): EmptyResult<NetworkError> {
+        return seravianChatBotDataSource.uploadUserVoice(UploadVoiceRequest(capturedVoice, chatId))
+    }
+
+    override suspend fun receiveAIAudioResponse(callback: (NetworkResult<Audio, NetworkError>) -> Unit) {
+        seravianChatBotDataSource.receiveAIAudioReadyResponse { aiAudioReadyResponse ->
+            lastNotifiedAudio = aiAudioReadyResponse.aiAudioId
+            val fetchingResult = downloadAudio(
+                aiAudioReadyResponse.extractFetchRequest()
+            )
+            callback(fetchingResult)
+        }
+    }
+
+    override suspend fun fetchAIAudio(
+        fetchRequest: FetchAIAudioRequest
+    ): NetworkResult<Audio, NetworkError> {
+        when {
+            fetchRequest.aiAudioId == lastNotifiedAudio && downloadExecution != null -> {
+                return NetworkResult.Error(NetworkError(ErrorType.TOO_MANY_REQUESTS))
+            }
+            fetchRequest.aiAudioId != lastNotifiedAudio && downloadExecution == null -> {
+                lastNotifiedAudio = fetchRequest.aiAudioId
+                return downloadAudio(fetchRequest)
+            }
+            fetchRequest.aiAudioId == lastNotifiedAudio && downloadExecution == null -> {
+                lastNotifiedAudio = fetchRequest.aiAudioId
+                return downloadAudio(fetchRequest)
+            }
+            else -> { return NetworkResult.Error(NetworkError(ErrorType.UNKNOWN_ERROR)) }
+        }
+    }
+
+    private suspend fun downloadAudio(
+        fetchRequest: FetchAIAudioRequest
+    ): NetworkResult<Audio, NetworkError> {
+        if (!audioDownloadLock.tryLock()) {
+            downloadExecution?.let {
+                println("Already Fetching Audio")
+                return NetworkResult.Error(NetworkError(ErrorType.TOO_MANY_REQUESTS))
+            }
+        }
+
+        return try {
+            val deferred = CoroutineScope(Dispatchers.IO).async {
+                seravianChatBotDataSource.fetchAIAudioResponse(fetchRequest).map { response ->
+                    response.extractAudio(fetchRequest.aiAudioId)
+                }
+            }
+            downloadExecution = deferred
+
+            deferred.await()
+        } finally {
+            audioDownloadLock.unlock()
+            downloadExecution = null
+        }
+    }
+}
