+++ chat/feat-chat/src/main/java/com/seravian/feat_chat/domain/repository/ChatRepository.kt
@@ -1,17 +1,10 @@
 package com.seravian.feat_chat.domain.repository
 
-import com.greenvenom.core_network.data.ConnectionStatus
-import com.greenvenom.core_network.data.EmptyResult
 import com.greenvenom.core_network.data.NetworkError
 import com.greenvenom.core_network.data.NetworkResult
 import com.seravian.core_chat.data.dto.request.ClientRequest
-import com.seravian.core_chat.data.dto.request.CreateChatRequest
-import com.seravian.core_chat.data.dto.request.DeleteChatRequest
-import com.seravian.core_chat.data.dto.request.EditChatRequest
 import com.seravian.core_chat.data.dto.request.FetchAIAudioRequest
 import com.seravian.core_chat.data.dto.request.GetChatMessagesRequest
-import com.seravian.core_chat.data.dto.request.JoinChatRequest
-import com.seravian.core_chat.data.dto.respose.AIAudioResponse
 import com.seravian.core_chat.data.dto.respose.ConfirmedMessageResponse
 import com.seravian.core_chat.domain.models.Audio
 import com.seravian.core_chat.domain.models.Chat
@@ -19,29 +12,11 @@ import com.seravian.core_chat.domain.models.Message
 import kotlinx.coroutines.flow.Flow
 
 interface ChatRepository {
-    suspend fun createChat(createChatRequest: CreateChatRequest): NetworkResult<Chat, NetworkError>
-
-    suspend fun updateChat(editChatRequest: EditChatRequest): NetworkResult<Chat, NetworkError>
-
-    suspend fun deleteChat(deleteChatRequest: DeleteChatRequest): EmptyResult<NetworkError>
-
-    fun getChats(): Flow<NetworkResult<List<Chat>, NetworkError>>
-
     fun getChatMessages(
         getChatMessagesRequest: GetChatMessagesRequest
     ): Flow<NetworkResult<Pair<Chat, List<Message>>, NetworkError>>
 
     suspend fun insertConfirmedMessage(message: Message)
-
-    suspend fun sendCapturedVoice(capturedVoice: ByteArray): EmptyResult<NetworkError>
-
-    suspend fun startConnection()
-
-    suspend fun stopConnection()
-
-    fun getSignalRConnectionStatus(): Flow<ConnectionStatus>
-
-    suspend fun joinChat(joinChatRequest: JoinChatRequest)
 
     suspend fun sendRequest(clientRequest: ClientRequest)
 
@@ -50,10 +25,4 @@ interface ChatRepository {
     fun receiveAIResponse()
 
     fun receiveMessageConfirmation(callback: suspend (ConfirmedMessageResponse) -> Unit)
-
-    suspend fun receiveAIAudioResponse(callback: (NetworkResult<Audio, NetworkError>) -> Unit)
-
-    suspend fun fetchAIAudio(
-        fetchRequest: FetchAIAudioRequest
-    ): NetworkResult<Audio, NetworkError>
 }
