+++ chat/feat-chat/src/main/java/com/seravian/feat_chat/domain/repository/ChatsListRepository.kt
@@ -1,0 +1,20 @@
+package com.seravian.feat_chat.domain.repository
+
+import com.greenvenom.core_network.data.EmptyResult
+import com.greenvenom.core_network.data.NetworkError
+import com.greenvenom.core_network.data.NetworkResult
+import com.seravian.core_chat.data.dto.request.CreateChatRequest
+import com.seravian.core_chat.data.dto.request.DeleteChatRequest
+import com.seravian.core_chat.data.dto.request.EditChatRequest
+import com.seravian.core_chat.domain.models.Chat
+import kotlinx.coroutines.flow.Flow
+
+interface ChatsListRepository {
+    suspend fun createChat(createChatRequest: CreateChatRequest): NetworkResult<Chat, NetworkError>
+
+    suspend fun updateChat(editChatRequest: EditChatRequest): NetworkResult<Chat, NetworkError>
+
+    suspend fun deleteChat(deleteChatRequest: DeleteChatRequest): EmptyResult<NetworkError>
+
+    fun getChats(): Flow<NetworkResult<List<Chat>, NetworkError>>
+}
